refactor: re-locate path helpers
