+++ src/common/storage.ts
@@ -1,5 +1,5 @@
 import { readFileSync, writeFileSync } from "fs";
-import { relative } from "./path";
+import { relative } from "../system/path";
 
 const storePath = relative("../.cli.cache");
