+++ src/package-manager.ts
@@ -5,7 +5,7 @@ import { Readable } from "node:stream";
 import semver from "semver";
 import tar from "tar";
 import { dependencies, version } from "../package.json";
-import { relative } from "./common/path";
+import { relative } from "./system/path";
 import { moveSync } from "./system/fs";
 
 /**
